code-samples/calendar/python/delete_calendar/pyproject.toml
[tool.poetry]
name = "delete-calendar"
version = "0.1.0"
description = "Delete calendar events using Nylas Python SDK"
authors = ["wiseaidev <oss@wiseai.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nylas = "^5.14.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.10.0"
pytest = "^7.4.2"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
delete_calendar = "delete_calendar.__main__:main"

[virtualenvs]
in-project = true
create = true